pyproject.toml
[pyproject.toml]
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "selenium-wrapper"
version = "0.1.0"
description = "USIT Support LLC - Python - Selenium Wrapper"
authors = [{name = "Romulo Conceicao", email = "romulo.conceicao@usitsupport.net"}]
license = {text = "MIT License"}
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "selenium>=4.0.0",
    "pyyaml>=6.0.0",
    "faker>=19.0.0",
]

[project.optional-dependencies]
testing = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "pytest-selenium>=4.1.1"
]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
    "pre-commit>=3.4.0"
]

[project.urls]
Homepage = "https://github.com/usit-support/selenium-wrapper"
Repository = "https://github.com/usit-support/selenium-wrapper.git"
"Bug Tracker" = "https://github.com/usit-support/selenium-wrapper/issues"
"Package Registry" = "https://github.com/orgs/usit-support/packages"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["selenium_wrapper*"]
exclude = ["tests*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test"]
python_functions = ["test_*"]
addopts = "-v --cov=selenium-wrapper"

[tool.coverage.run]
source = ["selenium-wrapper"]
omit = ["tests/*"]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.flake8]
max-line-length = 88
extend-ignore = "E203"
exclude = [".git", "__pycache__", "build", "dist"]

[tool.ruff]
line-length = 88
target-version = "py312"
select = ["E", "F", "B", "I", "W"]
ignore = ["E203"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]